backend/requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine learning
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
joblib>=1.3.0
sqlalchemy>=2.0.0
